Added: Left 4 Gore
The international version of Left 4 Dead 2 is banned in Australia and Germany, so games purchased in those countries are censored. Dismemberment, blood spatter, corpses, burning and more are all disabled in both versions. Left 4 Gore is a memory patcher which adds the gore back into these crippled versions of Left 4 Dead 2.